fixed segfault in combined when domain unloads
The issue was related to registration of CombinedAppHost transport. 
It was registered in the listener in main domain. The reference to the transport was
created in the constructor, we've got the unmanaged thunks to it, then created 
AppHostTransport when cretated Host. And this was wrong, because when domain 
unloads it unload all jit-method to the types are located in the domain and 
which are not referenced anymore.  And the CombinedAppHost was the type, that only
referenced from the domain.